dashboard/requirements.txt
pandas
streamlit 
psycopg2
altair 
python-dotenv 